Initial Commit
MQTT_Publish contains some code that connects to a mosquitto server and
publishes the temperature measured on a DS18B20. The code then holds a
delay and pushes an update every 10s. The whole thing may or may not
work depending on whether the connection remains open or not. A version
very similar to this (but without the ESP.deepsleep() on line 96 worked
for 10hours without failure in the initial test.