Wifi connection works, broker connection fails
WiFi network connection starts up with the system, and remains up (ping
proves this).
Added MQTT stuff, but there’s nothing working here. Issue seems to be
that it doesn’t connect to the broker (or fails). Nothing shows on
mosquito_sub on the broker end either, so it’s a double-ended fail.

Counter increments (slowly) since there’s a delay incurred with
connecting to the broker, and then failing.