Config server, and sleep
Configuration web server that opens a soft access point, and offers a
webpage that allows configuration of the SSID, password for the main
wifi network, and IP address of the MQTT broker.

Needs some tidying up of the page, and proper handler that returns some
sort of configuration confirmation, etc.

Once correctly connected, sends the temperature reading from the
DS18b20, and then sleeps for 5 minutes, ad infinitum.